Adding files via upload
I was not happy with the way I wrote the code for homebrew-passwordmanager and figured out it was better just to start over again and create a whole new repository for three main reasons: 1. the other code was really written badly (at least for me), it was hard to mantain. 2. I want to keep the other repo for remembering myself of my progress, and github is the best place to do that. 3. I still want to leave it public as it could be useful to someone (not very likely but still)